[4130] Fixed: load timed quest timer as time_t value. This fix possible unexpected timed quest fail at character load. Fixed: restore support build mangos at Unix/Linux in case build_dir != src_dir Fixed: lot warnings output at compile using GCC.

+++ src/game/ItemHandler.cpp
@@ -625,7 +625,7 @@ void WorldSession::SendListInventory( uint64 vendorguid )
         }
     }
 
-    if ( count == 0 || data.size() != 8 + 1 + count * 7 * 4 )
+    if ( count == 0 || data.size() != 8 + 1 + size_t(count) * 7 * 4 )
         return;
 
     data.put<uint8>(8, count);
